Optimized code structure and documentation
